requirements.txt
alembic==1.16.1
async-timeout==5.0.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
rich==13.9.4
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
